requirements.txt
discord.py>=2.7
mcp>=1.27
aiohttp>=3.13
